Update README with installation instructions
Also changes around some styling and adds a License section (though deliberately doesn't add a `LICENSE` file - GitHub should not be recognizing this project as MIT Licensed as the vast majority of the bytes are not subject to that license). Without any mention of a license potential contributors would be allowed to fork and view under the GitHub ToS, but technically making changes to the code they've forked is not covered.

A future commit should add a `CREDITS.md` file linked to from a `Credits` section in this README detailing those who have contributed their work to this repository.